Add files via upload
Presenting our improved version of a CAPTCHA solver. This one is addressed to blind users and helps them surf the Internet without any assistance. The accessible tool makes the difficult visual challenge, an intuitive one through the use of sound so that the users can hear the CAPTCHA prompts and complete it as a response.

Audio instructions if intricate parts are available in visual Captcha components are transformed into high and clear audios for use by the individuals performing the interested challenges even without the ability to see the parts.

Interface: Even a kid can use this interface, it is that simple and provides the best experience for every user, irrespective of their skills.

Multilanguage: Such content will be in audio format and screened in various languages accommodating almost every user and ensuring no prejudice while using the internet.

Real-Time processing: Solutions are provided fast and effectively analyzed which helps reduce the wastage of time and irritation especially while accessing sites with CAPTCHA.

Attention to Security: All measures are taken to prevent any data breaches and not a single trace of the user is kept or shared which is a #1 priority issue.

Such a CAPTCHA solver improves usability and goes a step further with embracing accessibility as it allows digital content for the visually impaired. Let us help each other make the internet better for everyone!